+++ packages/static-hosting/index.ts
@@ -1,9 +1,4 @@
-// import { StaticHosting, StaticHostingProps } from "./lib/static-hosting";
-// import { CSP } from "./types/csp";
-
-// export { StaticHosting, StaticHostingProps, CSP };
-
-import { StaticHosting, StaticHostingProps } from "./lib/old-static-hosting";
+import { StaticHosting, StaticHostingProps } from "./lib/static-hosting";
 import { CSP } from "./types/csp";
 
 export { StaticHosting, StaticHostingProps, CSP };
